backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==39.0.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.14.0
django-phonenumber-field==7.0.2
django-related-models==0.1.1
django-simple-history==3.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-writable-nested==0.7.0
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
Flask==2.2.2
fqdn==1.5.1
frozenlist==1.3.3
h11==0.14.0
heyoo==0.0.8
hyperlink==21.0.0
idna==3.4
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter-server-proxy==3.2.2
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.4
MouseInfo==0.1.3
multidict==6.0.4
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
outcome==1.2.0
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
phonenumbers==8.13.10
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.2.18
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.14.0
PyJWT==2.6.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
pytweening==1.0.4
pytz==2022.7.1
pywhatkit==5.4
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-toolbelt==0.9.1
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
selenium==4.7.2
Send2Trash==1.8.0
simpervisor==0.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
trio==0.22.0
trio-websocket==0.9.2
txaio==23.1.1
tzdata==2022.7
uri-template==1.2.0
urllib3==1.26.13
virtualenv==20.17.1
virtualenv-clone==0.5.7
vpython==7.6.4
wcwidth==0.2.6
webcolors==1.12
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.8.2
